+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Issue [#6](https://github.com/m-marini/wheellycam/issues/6): Add Add camera image size
 - Issue [#7](https://github.com/m-marini/wheellycam/issues/7): Add Add qr message
 
 ## [0.1.0] 2024-08-30